Reorganise prusti-contracts crates
