--- prusti-contracts/prusti-contracts-impl/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "prusti-contracts-impl"
-version = "0.1.0"
-authors = ["Vytautas Astrauskas <vastrauskas@gmail.com>"]
-edition = "2021"
-
-[lib]
-proc-macro = true
