+++ prusti-contracts/prusti-contracts-internal/Cargo.toml
@@ -8,5 +8,10 @@ edition = "2021"
 proc-macro = true
 
 [dependencies]
-prusti-specs = { path = "../prusti-specs" }
-proc-macro2 = "1.0"
+prusti-specs = { path = "../prusti-specs", optional = true }
+proc-macro2 = { version = "1.0", optional = true }
+
+[features]
+# Are we being compiled by Prusti and should include dependency on
+# prusti-specs and proc-macro2?
+prusti = ["dep:prusti-specs", "dep:proc-macro2"]
