+++ prusti-contracts/prusti-contracts/Cargo.toml
@@ -5,13 +5,11 @@ authors = ["Vytautas Astrauskas <vastrauskas@gmail.com>"]
 edition = "2021"
 
 [dependencies]
-prusti-contracts-impl = { path = "../prusti-contracts-impl" }
-prusti-contracts-internal = { path = "../prusti-contracts-internal", optional = true }
+prusti-contracts-internal = { path = "../prusti-contracts-internal" }
 
 [dev-dependencies]
 trybuild = "1.0"
 
 [features]
-# Are we being compiled by Prusti and should include dependency on
-# prusti-contracts-internal?
-prusti = ["prusti-contracts-internal"]
+# Are we being compiled by Prusti?
+prusti = ["prusti-contracts-internal/prusti"]
