+++ prusti-contracts/prusti-contracts/src/lib.rs
@@ -1,43 +1,64 @@
 #![no_std]
 
+/// A macro for writing a precondition on a function.
+pub use prusti_contracts_internal::requires;
+
+/// A macro for writing a postcondition on a function.
+pub use prusti_contracts_internal::ensures;
+
+/// A macro for writing a pledge on a function.
+pub use prusti_contracts_internal::after_expiry;
+
+/// A macro for writing a two-state pledge on a function.
+pub use prusti_contracts_internal::assert_on_expiry;
+
+/// A macro for marking a function as pure.
+pub use prusti_contracts_internal::pure;
+
+/// A macro for marking a function as trusted.
+pub use prusti_contracts_internal::trusted;
+
+/// A macro for type invariants.
+pub use prusti_contracts_internal::invariant;
+
+/// A macro for writing a loop body invariant.
+pub use prusti_contracts_internal::body_invariant;
+
+/// A macro for writing assertions using the full prusti specifications
+pub use prusti_contracts_internal::prusti_assert;
+
+/// A macro for writing assumptions using prusti syntax
+pub use prusti_contracts_internal::prusti_assume;
+
+/// A macro for impl blocks that refine trait specifications.
+pub use prusti_contracts_internal::refine_trait_spec;
+
+/// A macro for specifying external functions.
+pub use prusti_contracts_internal::extern_spec;
+
+/// A macro for defining a predicate using prusti expression syntax instead
+/// of just Rust expressions.
+pub use prusti_contracts_internal::predicate;
+
+/// Macro for creating type models.
+pub use prusti_contracts_internal::model;
+
+/// A macro to add trait bounds on a generic type parameter and specifications
+/// which are active only when these bounds are satisfied for a call.
+pub use prusti_contracts_internal::ghost_constraint;
+
+/// A macro for defining ghost blocks which will be left in for verification
+/// but omitted during compilation.
+pub use prusti_contracts_internal::ghost;
+
 #[cfg(not(feature = "prusti"))]
 mod private {
     use core::marker::PhantomData;
 
-    /// A macro for writing a precondition on a function.
-    pub use prusti_contracts_impl::requires;
-
-    /// A macro for writing a postcondition on a function.
-    pub use prusti_contracts_impl::ensures;
-
-    /// A macro for writing a pledge on a function.
-    pub use prusti_contracts_impl::after_expiry;
-
-    /// A macro for writing a two-state pledge on a function.
-    pub use prusti_contracts_impl::assert_on_expiry;
-
-    /// A macro for marking a function as pure.
-    pub use prusti_contracts_impl::pure;
-
-    /// A macro for marking a function as trusted.
-    pub use prusti_contracts_impl::trusted;
-
-    /// A macro for type invariants.
-    pub use prusti_contracts_impl::invariant;
-
-    /// A macro for writing a loop body invariant.
-    pub use prusti_contracts_impl::body_invariant;
-
-    /// A macro for writing assertions using the full prusti specifications
-    pub use prusti_contracts_impl::prusti_assert;
-
-    /// A macro for writing assumptions using prusti syntax
-    pub use prusti_contracts_impl::prusti_assume;
-
     /// A macro for defining a closure with a specification.
     /// Note: this is a declarative macro defined in this crate
     /// because declarative macros can't be exported from
-    /// the `prusti-contracts-impl` proc-macro crate.
+    /// the `prusti-contracts-internal` proc-macro crate.
     /// See <https://github.com/rust-lang/rust/issues/40090>.
     #[macro_export]
     macro_rules! closure {
@@ -48,23 +69,6 @@ mod private {
             $($tail)*
         };
     }
-
-    /// A macro for impl blocks that refine trait specifications.
-    pub use prusti_contracts_impl::refine_trait_spec;
-
-    /// A macro for specifying external functions.
-    pub use prusti_contracts_impl::extern_spec;
-
-    /// A macro for defining a predicate using prusti expression syntax instead
-    /// of just Rust expressions.
-    pub use prusti_contracts_impl::predicate;
-
-    /// Macro for creating type models.
-    pub use prusti_contracts_impl::model;
-
-    /// A macro to add trait bounds on a generic type parameter and specifications
-    /// which are active only when these bounds are satisfied for a call.
-    pub use prusti_contracts_impl::ghost_constraint;
 
     /// A sequence type
     #[non_exhaustive]
@@ -80,10 +84,6 @@ mod private {
         _key_phantom: PhantomData<K>,
         _val_phantom: PhantomData<V>,
     }
-
-    /// A macro for defining ghost blocks which will be left in for verification
-    /// but omitted during compilation.
-    pub use prusti_contracts_impl::ghost;
 
     /// a mathematical (unbounded) integer type
     /// it should not be constructed from running rust code, hence the private unit inside
@@ -103,58 +103,8 @@ pub mod core_spec;
 mod private {
     use core::{marker::PhantomData, ops::*};
 
-    /// A macro for writing a precondition on a function.
-    pub use prusti_contracts_internal::requires;
-
-    /// A macro for writing a postcondition on a function.
-    pub use prusti_contracts_internal::ensures;
-
-    /// A macro for writing a pledge on a function.
-    pub use prusti_contracts_internal::after_expiry;
-
-    /// A macro for writing a two-state pledge on a function.
-    pub use prusti_contracts_internal::assert_on_expiry;
-
-    /// A macro for marking a function as pure.
-    pub use prusti_contracts_internal::pure;
-
-    /// A macro for marking a function as trusted.
-    pub use prusti_contracts_internal::trusted;
-
-    /// A macro for type invariants.
-    pub use prusti_contracts_internal::invariant;
-
-    /// A macro for writing a loop body invariant.
-    pub use prusti_contracts_internal::body_invariant;
-
-    /// A macro for writing assertions using the full prusti specifications
-    pub use prusti_contracts_internal::prusti_assert;
-
-    /// A macro for writing assumptions using prusti syntax
-    pub use prusti_contracts_internal::prusti_assume;
-
     /// A macro for defining a closure with a specification.
     pub use prusti_contracts_internal::closure;
-
-    /// A macro for impl blocks that refine trait specifications.
-    pub use prusti_contracts_internal::refine_trait_spec;
-
-    /// A macro for specifying external functions.
-    pub use prusti_contracts_internal::extern_spec;
-
-    /// A macro for defining a predicate using prusti expression syntax instead
-    /// of just Rust expressions.
-    pub use prusti_contracts_internal::predicate;
-
-    /// Macro for creating type models.
-    pub use prusti_contracts_internal::model;
-
-    /// A macro to add trait bounds on a generic type parameter and specifications
-    /// which are active only when these bounds are satisfied for a call.
-    pub use prusti_contracts_internal::ghost_constraint;
-
-    ///A macro to custamize how a struct or enum should be printed in a counterexample
-    pub use prusti_contracts_internal::print_counterexample;
 
     pub fn prusti_set_union_active_field<T>(_arg: T) {
         unreachable!();
@@ -331,10 +281,6 @@ mod private {
             panic!()
         }
     }
-
-    /// A macro for defining ghost blocks which will be left in for verification
-    /// but omitted during compilation.
-    pub use prusti_contracts_internal::ghost;
 }
 
 /// This function is used to evaluate an expression in the context just
